# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MedineTech - ERP - Management Software API",
        "description": "Integration API Docs.",
        "contact": {
            "name": "Developers Team",
            "url": "https://medine.tech/",
            "email": "devs@medine.tech"
        },
        "version": "0.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Localhost Server"
        }
    ],
    "paths": {
...
        "/api/backoffice/{tenant}/accounting/accounting-centers/{id}": {
            "get": {
                "tags": [
                    "Backoffice - Accounting - Accounting Centers"
                ],
                "summary": "Retrieve an accounting center by ID",
                "operationId": "7e59caac36647f09a9cfb1fe04be0619",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Accounting Center ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Accounting Center found successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "code": {
                                            "type": "string",
                                            "example": "ACC-01"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Main Accounting Center"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Description of the Accounting Center"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "ACTIVE"
                                        },
                                        "parentId": {
                                            "type": "string",
                                            "example": "NULL",
                                            "nullable": true
                                        },
                                        "creatorId": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "updaterId": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "companyId": {
                                            "type": "string",
                                            "example": "COMPANY-UUID"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Accounting Center Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "title": {
                                            "type": "string",
                                            "example": "Accounting Center Not Found"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "detail": {
                                            "type": "string",
                                            "example": "Accounting Center with the given ID does not exist"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "title": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 403
                                        },
                                        "detail": {
                                            "type": "string",
                                            "example": "You do not have permission to view this resource."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "title": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "detail": {
                                            "type": "string",
                                            "example": "An Unexpected Error Occurred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Backoffice - Accounting - Accounting Centers"
                ],
                "summary": "Update an existing accounting center",
                "operationId": "637f3147b0942fee681350bac51b3e4a",
                "parameters": [
                    {
                        "name": "tenant",
                        "in": "path",
                        "description": "The tenant ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the accounting center to update",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "status"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Center Name"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated accounting center description",
                                        "nullable": true
                                    },
                                    "status": {
                                        "description": "Center status",
                                        "type": "string",
                                        "example": "INACTIVE"
                                    },
                                    "parent_id": {
                                        "description": "Parent center ID",
                                        "type": "string",
                                        "example": "123e4567-e89b-12d3-a456-426655440000",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Accounting center updated successfully"
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "title": {
                                            "type": "string",
                                            "example": "Validation Error"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "detail": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Accounting center not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "title": {
                                            "type": "string",
                                            "example": "Not Found"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "detail": {
                                            "type": "string",
                                            "example": "Accounting center not found with the provided ID."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "title": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 403
                                        },
                                        "detail": {
                                            "type": "string",
                                            "example": "You do not have permission to update this resource."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "title": {
                                            "type": "string",
                                            "example": "Error"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "detail": {
                                            "type": "string",
                                            "example": "An unexpected error occurred while processing your request."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}